backend/requirements.txt
fastapi[standard]
pydantic-settings
motor
redis 
bcrypt
pyjwt
python-dotenv
pytest
httpx
slowapi